+++ leilao-api/src/main/java/com/higotlino/leilao/service/EmpresaService.java
@@ -1,8 +1,10 @@
 package com.higotlino.leilao.service;
 
 import com.higotlino.leilao.entity.Empresa;
+import com.higotlino.leilao.exception.ResourceNotFoundException;
 import com.higotlino.leilao.repository.EmpresaRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -17,7 +19,8 @@ public class EmpresaService {
 
     @Transactional (readOnly = true)
     public Empresa getById(Long id){
-        return empresaRepository.findById(id).orElse(null);
+        return empresaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Empresa nao encontrada"));
     }
 
     @Transactional (readOnly = true)
@@ -28,16 +31,30 @@ public class EmpresaService {
     @Transactional
     public Empresa create(Empresa empresa) {
         empresa.setPassword(passwordEncoder.encode(empresa.getPassword()));
+        if (empresaRepository.existsEmpresaByCnpj(empresa.getCnpj()))
+            throw new DataIntegrityViolationException("Empresa ja cadastrada com este CNPJ");
+
+        if (empresaRepository.existsEmpresaByUsuario(empresa.getUsuario()))
+            throw new DataIntegrityViolationException("Empresa ja cadastrada com este usuario");
+
         return empresaRepository.save(empresa);
     }
 
     @Transactional
     public Empresa update(Empresa empresa) {
+        if (empresaRepository.existsEmpresaByCnpj(empresa.getCnpj()))
+            throw new DataIntegrityViolationException("Empresa ja cadastrada com este CNPJ");
+
+        if (empresaRepository.existsEmpresaByUsuario(empresa.getUsuario()))
+            throw new DataIntegrityViolationException("Empresa ja cadastrada com este usuario");
+
         return empresaRepository.save(empresa);
     }
 
     @Transactional
     public void delete(Long id) {
+        if (empresaRepository.findById(id).isEmpty())
+            throw new ResourceNotFoundException("Empresa nao encontrada");
         empresaRepository.deleteById(id);
     }
 }
